\name{StatComp20002-package}
\alias{StatComp20002-package}
\alias{StatComp20002}
\docType{package}
\title{
  The final project of the course "Statistical Computing"
}
\description{
  This package provides all functions from my own homeworks and two other functions.
}
\details{
  This package provides all functions from my own homeworks and two functions (\code{rcauchyMH} for generating random samples from Cauchy distribution by MH sampler, \code{SampleChain} for generateing a Metropolis chain for multinomial distribution with random walk proposal distribution and starting value start).
}
\author{
Hongrui Cai

Maintainer: Hongrui Cai <rainbow_rui8@163.com>
}
\keyword{ package }
