\name{StatComp20002-package}
\alias{StatComp20002-package}
\alias{StatComp20002}
\docType{package}
\title{
  The final project of the course "Statistical Computing"
}
\description{
  This package provides two functions for benchmarking R and Rcpp, and all functions for my own homeworks.
}
\details{
  This package implements two functions and all functions for my homeworks.
}
\author{
Hongrui Cai

Maintainer: Hongrui Cai <rainbow_rui8@163.com>
}
\keyword{ package }